requirements.txt
asgiref==3.11.0
distlib==0.4.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.20.1
gunicorn==23.0.0
packaging==25.0
platformdirs==4.5.1
sqlparse==0.5.5
stevedore==5.6.0
tzdata==2025.3
virtualenv==20.35.4
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
virtualenvwrapper-win==1.2.7
